etcd delete handler added
